Revert "feat(tui): passive active-monitor/cron visualization panel (#675)"
This reverts commit a8d0d5f4e1ff577f5c1325f58f57830270109127.
