+++ packages/coding-agent/src/config/keybindings.ts
@@ -39,8 +39,6 @@ interface AppKeybindings {
 	"app.session.resume": true;
 	"app.session.observe": true;
 	"app.jobs.open": true;
-	"app.jobs.expand": true;
-	"app.jobs.collapse": true;
 	"app.session.togglePath": true;
 	"app.session.toggleSort": true;
 	"app.session.rename": true;
@@ -156,14 +154,6 @@ export const KEYBINDINGS = {
 	"app.jobs.open": {
 		defaultKeys: "alt+j",
 		description: "Open monitor/cron jobs overlay",
-	},
-	"app.jobs.expand": {
-		defaultKeys: "ctrl+up",
-		description: "Expand the active monitor/cron panel (or scroll up when expanded)",
-	},
-	"app.jobs.collapse": {
-		defaultKeys: "ctrl+down",
-		description: "Collapse the active monitor/cron panel (or scroll down when expanded)",
 	},
 	"app.session.togglePath": {
 		defaultKeys: "ctrl+p",
